requirements.txt
annotated-types==0.5.0
anyio==4.0.0
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
contourpy==1.1.0
copier==8.3.0
cryptography==41.0.3
cycler==0.11.0
decorator==5.1.1
Deprecated==1.2.14
dunamai==1.18.0
et-xmlfile==1.1.0
Flask==2.3.3
flet==0.10.1
flet-core==0.10.1
flet-route==0.3.2
flet-runtime==0.10.1
fonttools==4.42.1
funcy==2.0
gcloud==0.18.3
googleapis-common-protos==1.60.0
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jinja2-ansible-filters==1.3.2
joblib==1.3.2
jwcrypto==1.5.0
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.7.2
numpy==1.25.2
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.1
pandas==2.1.0
pathspec==0.11.2
patsy==0.5.3
Pillow==10.0.0
plumbum==1.8.2
prompt-toolkit==3.0.36
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.16.1
pyparsing==3.0.9
pypng==0.20220715.0
Pyrebase4==4.7.1
python-dateutil==2.8.2
python-jwt==4.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyyaml-include==1.3.1
qrcode==7.4.2
questionary==2.0.1
repath==0.9.0
requests==2.29.0
requests-toolbelt==0.10.1
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.2
six==1.16.0
sklearn==0.0.post9
sniffio==1.3.0
statsmodels==0.14.0
threadpoolctl==3.2.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
watchdog==3.0.0
wcwidth==0.2.6
websocket-client==1.6.2
websockets==11.0.3
Werkzeug==2.3.7
wrapt==1.15.0
